Fix husky git hook snippet in local setup
The snippet to setup husky to run cz when `git commit` is run assumes the developer has setup the command `git cz` globally.
This throws an error: `git: 'cz' is not a git command. See 'git --help'.`

The update uses the same reference to the locally installed cz as the git hook example above it.